build: resolve compile warnings on Solaris/OmniOS
lib/threads_pool.cpp: In function 'void* tpol_thrwork(void*)':
lib/threads_pool.cpp:200:25: note: 'snprintf' output between 10 and
28 bytes into a destination of size 16
  200 |                 snprintf(buf, std::size(buf), "ep_pool/%ld", pdata->nid;
lib/mapi/rule_actions.cpp: In function 'FORWARDDELEGATE_ACTION*
forwarddelegate_action_dup(const FORWARDDELEGATE_ACTION*)':
lib/mapi/rule_actions.cpp:148:27: warning: comparison of integer
expressions of different signedness: 'int' and 'const uint32_t' {aka
'const unsigned int'} [-Wsign-compare]
  148 |         for (int i = 0; i < paction->count; ++i) {

+++ lib/mapi/rule_actions.cpp
@@ -145,7 +145,7 @@ static FORWARDDELEGATE_ACTION* forwarddelegate_action_dup(
 		free(pblock);
 		return NULL;
 	}
-	for (int i = 0; i < paction->count; ++i) {
+	for (size_t i = 0; i < paction->count; ++i) {
 		if (recipient_block_dup_internal(&paction->pblock[i], &pblock->pblock[i]))
 			continue;
 		for (i -= 1; i >= 0; --i)
